Moved Storage rules integration tests under scripts/
