+++ scripts/storage-emulator-integration/rules/runtime.test.ts
@@ -1,19 +1,16 @@
 import {
   RulesetVerificationOpts,
   StorageRulesRuntime,
-} from "../../../../emulator/storage/rules/runtime";
+} from "../../../src/emulator/storage/rules/runtime";
 import { expect } from "chai";
-import { StorageRulesFiles } from "../../fixtures";
+import { StorageRulesFiles } from "../../emulator-tests/fixtures";
 import * as jwt from "jsonwebtoken";
-import { EmulatorLogger } from "../../../../emulator/emulatorLogger";
-import { ExpressionValue } from "../../../../emulator/storage/rules/expressionValue";
-import { RulesetOperationMethod } from "../../../../emulator/storage/rules/types";
-import {
-  downloadIfNecessary,
-  getDownloadDetails,
-} from "../../../../emulator/downloadableEmulators";
-import { Emulators } from "../../../../emulator/types";
-import { RulesResourceMetadata } from "../../../../emulator/storage/metadata";
+import { EmulatorLogger } from "../../../src/emulator/emulatorLogger";
+import { ExpressionValue } from "../../../src/emulator/storage/rules/expressionValue";
+import { RulesetOperationMethod } from "../../../src/emulator/storage/rules/types";
+import { downloadIfNecessary } from "../../../src/emulator/downloadableEmulators";
+import { Emulators } from "../../../src/emulator/types";
+import { RulesResourceMetadata } from "../../../src/emulator/storage/metadata";
 
 const TOKENS = {
   signedInUser: jwt.sign(
@@ -46,7 +43,7 @@ function createFakeResourceMetadata(params: {
   };
 }
 
-describe.skip("Storage Rules Runtime", function () {
+describe("Storage Rules Runtime", function () {
   let runtime: StorageRulesRuntime;
 
   // eslint-disable-next-line @typescript-eslint/no-invalid-this
