+++ scripts/storage-emulator-integration/run.sh
@@ -7,6 +7,8 @@ set -e # Immediately exit on failure
 # Prepare the storage emulator rules runtime
 firebase setup:emulators:storage
 
+mocha scripts/storage-emulator-integration/rules/*.test.ts
+
 mocha \
   --require ts-node/register \
   --require source-map-support/register \
